django_react_news/drf/serializers.py
from rest_framework import serializers
from .models import News

class NewsSerializers(serializers.ModelSerializer):
    user = serializers.HiddenField(default=serializers.CurrentUserDefault())
    class Meta:
        model = News
        fields = '__all__'